Modularize routes, and set up template to use in future route creation
